requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-environ>=0.11
django-cors-headers>=4.3
django-cloudinary-storage>=0.3
cloudinary>=1.36
google-generativeai>=0.5
PyMySQL>=1.1           # Pure Python MySQL driver (no system libs needed)
whitenoise>=6.6
gunicorn>=21.2
Pillow>=10.0
